All copies of NightShade die when the main one dies Added boss music to DraugrBoss and NightShade (placeholder) and discs - Commented out frenzied shade disc for now Added config lines for the armor of all bosses Buffed HP of all bosses Buffed armor of bosses Removed DARKNESS effect off of NightShade's attack ReturningKnight only summons if no healer is alive and will no longer summon healers under 50% health - Obliterate will summon double amount of Remnants if it kills something (not only undead anymore) - Added config line for summon cooldown on ReturningKnight Fixed a bug preventing DayStalker from using OVERHEAT attack Fixed a bug preventing NightProwler from using BLADES REACH and RIPPLE FANG attacks Fixed bugs with doConsumerOnCircle method Added hurt sounds to NightProwler and DayStalker Excalibur Sonic boom attack won't auto target armor stands LeviathanAxe and Mjolnir won't collide with owner Added config lines for other entities Increased some entities' health and armor Can now use lost souls in the common tag to heal Soulmass Added config line for FreyrSwordEntity health and armor

+++ src/main/java/net/soulsweaponry/util/ModifyDamageUtil.java
@@ -61,7 +61,7 @@ public class ModifyDamageUtil {
             int amplifier = entity.getStatusEffect(EffectRegistry.BLIGHT).getAmplifier() + 1; // ln(0) does not end well!
             int armorValue = entity.getArmor();
             // Original value increases in % based on f(x) = (ln(x) * y) / 8, where x is the amplifier level, y is the armor of the target.
-            double increase = newAmount * (((Math.log(amplifier) * armorValue) / 6f) / 10f);
+            float increase = (float) (newAmount * (((Math.log(amplifier) * armorValue) / 6f) / 10f));
             newAmount += increase;
         }
         if (source.getSource() instanceof TrueDamageArrow projectile) {
